Rename `MaterializationConfig` to `Materialization`
